Test just the pretty maker in the main images file.
